4. Data Logging and Service Tests
Added system to log actions performed against a user including viewing the user, editing the user, deleting the user and adding a new user including a new controller and new views. Refactored a lot of code to make things work with the logging. Began adding user service tests. Installed the full version of Xunit so that Assert could be used.